% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gate_for_biocasia2021}
\alias{gate_for_biocasia2021}
\title{gate_for_biocasia2021}
\format{
A list containing x,y coordinates for one gate
}
\usage{
data(gate_for_biocasia2021)
}
\description{
Coordinates for a gate interactively drawn using tidygate
}
\keyword{datasets}
