% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_mito_info_all_datasets}
\alias{pbmc_mito_info_all_datasets}
\title{pbmc_mito_info_all_datasets}
\format{
A SingleCellExperiment object.
}
\usage{
data(pbmc_mito_info_all_datasets)
}
\description{
A SingleCellExperiment dataset of single cell RNA sequencing data. Used for one of the exercises
}
\keyword{datasets}
